requirements.txt
Django==2.2.19
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
PyJWT==1.6.1
pytz==2018.3
